refactor: separate code about DB
